# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products/menu": {
      "get": {
        "operationId": "Products_getMenu",
        "summary": "Get active POS menu grouped by category",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/products": {
      "post": {
        "operationId": "Products_create",
        "summary": "Create a new product",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Products_findAll",
        "summary": "Get all products with pagination and filters",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRINK",
                "TOPPING",
                "MATERIAL"
              ],
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProductResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "Products_findOne",
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "Products_update",
        "summary": "Update a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "Products_remove",
        "summary": "Soft delete a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/recipes/{drinkId}": {
      "get": {
        "operationId": "Recipes_getRecipe",
        "summary": "Get recipe for a specific drink",
        "parameters": [
          {
            "name": "drinkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "put": {
        "operationId": "Recipes_setRecipe",
        "summary": "Set or update recipe for a specific drink",
        "parameters": [
          {
            "name": "drinkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/promotions": {
      "post": {
        "operationId": "Promotions_create",
        "summary": "Create a new promotion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Promotions_findAll",
        "summary": "Get all promotions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/promotions/active": {
      "get": {
        "operationId": "Promotions_findActive",
        "summary": "Get all active promotions within valid date range",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/promotions/{id}": {
      "get": {
        "operationId": "Promotions_findOne",
        "summary": "Get a promotion by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "Promotions_update",
        "summary": "Update a promotion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/promotions/{id}/deactivate": {
      "post": {
        "operationId": "Promotions_deactivate",
        "summary": "Deactivate a promotion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Promotions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TeaP ERP/POS API",
    "description": "API documentation for TeaP — Bubble Tea Chain ERP/POS Management System. Supports multi-branch operations, POS, inventory management, HR, and finance.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "CategoryReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ProductSizeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priceAdj": {
            "type": "integer",
            "description": "Whole VND",
            "example": 5000
          }
        },
        "required": [
          "id",
          "productId",
          "name",
          "priceAdj"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DRINK",
              "TOPPING",
              "MATERIAL"
            ]
          },
          "basePrice": {
            "type": "integer",
            "description": "Whole VND",
            "example": 45000
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryReferenceDto"
              }
            ]
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductSizeResponseDto"
            }
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "type",
          "basePrice",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "MenuCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "sortOrder",
          "isActive",
          "products"
        ]
      },
      "MenuResponseDto": {
        "type": "object",
        "properties": {
          "drinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "toppings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuCategoryResponseDto"
            }
          }
        },
        "required": [
          "drinks",
          "toppings",
          "categories"
        ]
      },
      "ProductSizeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Size name, e.g., S, M, L"
          },
          "priceAdjustment": {
            "type": "integer",
            "minimum": -9999999999,
            "maximum": 9999999999,
            "description": "Price adjustment in whole VND"
          }
        },
        "required": [
          "name",
          "priceAdjustment"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "description": "Unique SKU for the product"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "type": {
            "type": "string",
            "enum": [
              "DRINK",
              "TOPPING",
              "MATERIAL"
            ],
            "description": "Type of product"
          },
          "basePrice": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Base price in whole VND",
            "example": 45000
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID"
          },
          "image": {
            "type": "string",
            "description": "Image URL"
          },
          "sizes": {
            "description": "Product sizes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductSizeDto"
            }
          }
        },
        "required": [
          "sku",
          "name",
          "type",
          "basePrice"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "description": "Unique SKU for the product"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "type": {
            "type": "string",
            "enum": [
              "DRINK",
              "TOPPING",
              "MATERIAL"
            ],
            "description": "Type of product"
          },
          "basePrice": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Base price in whole VND",
            "example": 45000
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID"
          },
          "image": {
            "type": "string",
            "description": "Image URL"
          },
          "sizes": {
            "description": "Product sizes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductSizeDto"
            }
          }
        }
      },
      "RecipeMaterialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DRINK",
              "TOPPING",
              "MATERIAL"
            ]
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "type"
        ]
      },
      "RecipeSizeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RecipeItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "drinkId": {
            "type": "string"
          },
          "materialId": {
            "type": "string"
          },
          "sizeId": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "string",
            "example": "25.5",
            "pattern": "^-?\\d+(?:\\.\\d+)?$",
            "description": "Exact decimal quantity; parse explicitly when calculating"
          },
          "unit": {
            "type": "string"
          },
          "material": {
            "$ref": "#/components/schemas/RecipeMaterialResponseDto"
          },
          "size": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RecipeSizeResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "drinkId",
          "materialId",
          "quantity",
          "unit"
        ]
      },
      "RecipeItemDto": {
        "type": "object",
        "properties": {
          "materialId": {
            "type": "string",
            "description": "ID of the material product"
          },
          "sizeId": {
            "type": "string",
            "description": "ID of the product size (if specific to a size)"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the material used"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measurement (e.g., ml, g, pcs)"
          }
        },
        "required": [
          "materialId",
          "quantity",
          "unit"
        ]
      },
      "UpdateRecipeDto": {
        "type": "object",
        "properties": {
          "drinkId": {
            "type": "string",
            "description": "ID of the drink product"
          },
          "items": {
            "description": "List of recipe items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeItemDto"
            }
          }
        },
        "required": [
          "drinkId",
          "items"
        ]
      },
...
      "CreatePromotionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the promotion"
          },
          "description": {
            "type": "string",
            "description": "Description of the promotion"
          },
          "type": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT",
              "BUY_X_GET_Y"
            ],
            "description": "Type of promotion"
          },
          "value": {
            "type": "number",
            "description": "Discount value (percentage or fixed amount)"
          },
          "minOrderValue": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Minimum order value required for promotion"
          },
          "maxDiscount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Maximum discount amount allowed"
          },
          "startDate": {
            "type": "string",
            "description": "Start date of the promotion"
          },
          "endDate": {
            "type": "string",
            "description": "End date of the promotion"
          }
        },
        "required": [
          "name",
          "type",
          "value",
          "startDate",
          "endDate"
        ]
      },
      "UpdatePromotionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the promotion"
          },
          "description": {
            "type": "string",
            "description": "Description of the promotion"
          },
          "type": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT",
              "BUY_X_GET_Y"
            ],
            "description": "Type of promotion"
          },
          "value": {
            "type": "number",
            "description": "Discount value (percentage or fixed amount)"
          },
          "minOrderValue": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Minimum order value required for promotion"
          },
          "maxDiscount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9999999999,
            "description": "Maximum discount amount allowed"
          },
          "startDate": {
            "type": "string",
            "description": "Start date of the promotion"
          },
          "endDate": {
            "type": "string",
            "description": "End date of the promotion"
          }
        }
      },
...
    }
  }
}